Added some casts for lint

+++ modules/src/flt_arith/b64_add.c
@@ -15,11 +15,11 @@ flt_b64_add(e1,e2)
 	int	carry;
 
 	/* add higher pair of 32 bits */
-	overflow = ucmp(0xFFFFFFFF - e1->flt_h_32, e2->flt_h_32) < 0;
+	overflow = ucmp((long)0xFFFFFFFF - e1->flt_h_32, e2->flt_h_32) < 0;
 	e1->flt_h_32 += e2->flt_h_32;
 
 	/* add lower pair of 32 bits */
-	carry = ucmp(0xFFFFFFFF - e1->flt_l_32, e2->flt_l_32) < 0;
+	carry = ucmp((long)0xFFFFFFFF - e1->flt_l_32, e2->flt_l_32) < 0;
 	e1->flt_l_32 += e2->flt_l_32;
 
 	if ((carry) && ((++e1->flt_h_32 &~0xFFFFFFFF) || e1->flt_h_32 == 0)) {
